+++ app/helpers/favourite_rooms_helper.rb
@@ -1,0 +1,2 @@
+module FavouriteRoomsHelper
+end
